Split tests to separate modules
